+++ lib/ui_blocs/places_cubit.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:places/domain_models/place.dart';
 import 'package:places/main.dart';
 import 'package:places/ui_commons/platform_detector.dart';
-import 'package:places/ui_common_widgets/widget_add_to_calendar_cuper_modal.dart';
+import 'package:places/ui_widgets_commons/widget_add_to_calendar_cuper_modal.dart';
 
 class PlacesCubit extends Cubit<VisitedAndFavoriteScreenState> {
   PlacesCubit(VisitedAndFavoriteScreenState initialState) : super(initialState);
