+++ lib/ui_commons/themes.dart
@@ -36,7 +36,7 @@ final lightTheme = ThemeData(
   colorScheme: ColorScheme.fromSwatch().copyWith(
     secondary: const Color(0xFF4CAF50),
   ),
-  splashColor: Colors.red, // цвет риппл эффекта??
+  splashColor: const Color(0x774CAF50), // цвет риппл эффекта??
 );
 
 final darkTheme = ThemeData(
@@ -67,5 +67,5 @@ final darkTheme = ThemeData(
   colorScheme: ColorScheme.fromSwatch().copyWith(
     secondary: const Color(0xFF4CAF50),
   ),
-  splashColor: Colors.red, // цвет риппл эффекта??
+  splashColor: const Color(0x774CAF50), // цвет риппл эффекта??
 );
