+++ lib/ui_interactors/place_interactor.dart
@@ -6,7 +6,7 @@ import 'package:places/domain_models/filter_item.dart';
 import 'package:places/ui_commons/platform_detector.dart';
 
 import 'package:places/domain_models/place.dart';
-import 'package:places/ui_common_widgets/widget_add_to_calendar_cuper_modal.dart';
+import 'package:places/ui_widgets_commons/widget_add_to_calendar_cuper_modal.dart';
 
 class PlaceInteractor with ChangeNotifier {
   // <singleton>
