+++ lib/ui_screens/add_place_screen/screen_add_place.dart
@@ -2,8 +2,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:places/my_app_and_routes.dart';
 import 'package:places/ui_screens/add_place_screen/widget_new_place_app_bar.dart';
-import 'package:places/ui_common_widgets/widget_bottom_button.dart';
-import 'package:places/ui_common_widgets/widget_textfield_clear_button.dart';
+import 'package:places/ui_widgets_commons/widget_bottom_button.dart';
+import 'package:places/ui_widgets_commons/widget_textfield_clear_button.dart';
 import 'package:places/ui_commons/platform_detector.dart';
 import 'package:places/ui_commons/ui_strings.dart';
 import 'package:provider/provider.dart';
@@ -12,7 +12,7 @@ import 'package:places/ui_commons/my_scroll_physics.dart';
 import 'package:places/ui_screens/add_place_screen/widget_small_category_header.dart';
 import 'package:places/ui_screens/add_place_screen/dialog_add_photo.dart';
 import 'package:places/ui_interactors/select_category_interactor.dart';
-import 'package:places/ui_common_widgets/widget_my_image.dart';
+import 'package:places/ui_widgets_commons/widget_my_image.dart';
 import 'package:places/ui_interactors/place_interactor.dart';
 
 /// Экран - Добавить место
