+++ lib/ui_screens/main_1_places_list/screen_main1_places_route.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:places/ui_screens/main_1_places_list/screen_main1_places_narrow.dart';
 import 'package:places/ui_screens/main_1_places_list/screen_main1_places_wide.dart';
 import 'package:places/ui_interactors/place_interactor.dart';
-import 'package:places/ui_common_widgets/widget_network_error.dart';
+import 'package:places/ui_widgets_commons/widget_network_error.dart';
 import 'package:places/domain_models/place.dart';
 
 /// Экран 1. Список мест.
