+++ lib/ui_screens/main_3_wished_and_seen/screen_main3_favorite_and_visited.dart
@@ -3,7 +3,7 @@ import 'package:places/domain_models/place.dart';
 
 import 'package:places/ui_screens/main_3_wished_and_seen/widget_tab_favorite.dart';
 import 'package:places/ui_screens/main_3_wished_and_seen/widget_tab_visited.dart';
-import 'package:places/ui_common_widgets/widget_my_header.dart';
+import 'package:places/ui_widgets_commons/widget_my_header.dart';
 
 /// Экран 3. Список избранных и посещенных мест.
 /// Третий из четырех главных экранов, доступных по нажатию на
