+++ lib/ui_screens/main_3_wished_and_seen/screen_main3_favorite_and_visited_route.dart
@@ -4,7 +4,7 @@ import 'package:places/domain_models/place.dart';
 import 'package:places/main.dart';
 import 'package:places/ui_screens/main_3_wished_and_seen/screen_main3_favorite_and_visited.dart';
 import 'package:places/ui_blocs/places_cubit.dart';
-import 'package:places/ui_common_widgets/widget_network_error.dart';
+import 'package:places/ui_widgets_commons/widget_network_error.dart';
 
 /// Экран 3. Список избранных и посещенных мест.
 /// Третий из четырех главных экранов, доступных по нажатию на
