+++ lib/ui_screens/main_3_wished_and_seen/widget_tab_favorite.dart
@@ -8,7 +8,7 @@ import 'package:provider/provider.dart';
 
 import 'package:places/ui_commons/enums.dart';
 import 'package:places/ui_commons/my_scroll_physics.dart';
-import 'package:places/ui_common_widgets/widget_place_card.dart';
+import 'package:places/ui_widgets_commons/widget_place_card.dart';
 import 'package:places/ui_screens/place_details_screen/screen_place_details.dart';
 
 import 'package:places/domain_models/place.dart';
@@ -53,23 +53,26 @@ class _WidgetTabFavoriteState extends State<WidgetTabFavorite> {
                 .map(
                   (i) => Column(
                     children: [
-                      WidgetPlaceCard(
-                        place: i.value,
-                        placeCardType: PlaceCardType.wished,
-                        onTap: () {
-                          onTapOnCard(i.value.id);
-                        },
-                        onAddToCalendar: () {
-                          context.read<PlacesCubit>().schedulePlace(
-                                context,
-                                i.value.id,
-                              );
-                        },
-                        onDeleteFromWished: () {
-                          context
-                              .read<PlacesCubit>()
-                              .removeFromFavorites(_listOfItems[i.key].id);
-                        },
+                      Padding(
+                        padding: const EdgeInsets.symmetric(vertical: 8),
+                        child: WidgetPlaceCard(
+                          place: i.value,
+                          placeCardType: PlaceCardType.wished,
+                          onTap: () {
+                            onTapOnCard(i.value.id);
+                          },
+                          onAddToCalendar: () {
+                            context.read<PlacesCubit>().schedulePlace(
+                                  context,
+                                  i.value.id,
+                                );
+                          },
+                          onDeleteFromWished: () {
+                            context
+                                .read<PlacesCubit>()
+                                .removeFromFavorites(_listOfItems[i.key].id);
+                          },
+                        ),
                       ),
                     ],
                   ),
