+++ lib/ui_screens/place_details_screen/screen_place_details.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import 'package:places/ui_commons/ui_strings.dart';
 import 'package:places/ui_commons/my_scroll_physics.dart';
 import 'package:places/ui_screens/place_details_screen/widget_back_button.dart';
-import 'package:places/ui_common_widgets/widget_my_image.dart';
+import 'package:places/ui_widgets_commons/widget_my_image.dart';
 
 import 'package:places/domain_models/place.dart';
