+++ lib/ui_screens/place_search_screen/widget_search_result.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:places/domain_models/place.dart';
-import 'package:places/ui_common_widgets/widget_place_card_for_search.dart';
+import 'package:places/ui_widgets_commons/widget_place_card_for_search.dart';
 import 'package:places/ui_interactors/search_interactor.dart';
 import 'package:provider/provider.dart';
