+++ lib/ui_widgets_commons/widget_network_error.dart
