+++ lib/ui_widgets_commons/widget_place_card_for_search.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:places/ui_screens/place_details_screen/screen_place_details.dart';
-import 'package:places/ui_common_widgets/widget_my_image.dart';
+import 'package:places/ui_widgets_commons/widget_my_image.dart';
 
 import 'package:places/domain_models/place.dart';
