+++ lib/ui_widgets_commons/widget_textfield_clear_button.dart
